Modifhy By TL

+++ GIT_NITProj01/src/main/java/com/nt/service/Arithmetic.java
@@ -1,7 +1,7 @@
 package com.nt.service;
 
 public class Arithmetic {
-	public int sum(int x,int y) {
+	public double sum(double x,double y) {
 		return x+y;
 	}
